requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lancedb==0.8.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
ollama==0.2.0
openai==1.30.5
overrides==7.7.0
packaging==24.0
pandas==2.2.2
phidata==2.4.17
py==1.11.0
pyarrow==15.0.0
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
pylance==0.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
requests==2.32.3
retry==0.9.2
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
